extracted generateUrls
